chore: breakup query helpers into pre- and post- processing
